Move the `get_column_values()` function from the `cleaning` module to the `info` module
